move save button to view dialog
